commented out empty footer added routes in a pp
