+++ client/src/components/Layout/Layout.jsx
@@ -8,7 +8,7 @@ const Layout = (props) => (
     <div className="layout-children">
       {props.children}
     </div>
-    <Footer />
+    {/* <Footer /> */}
   </div>
 )
